Work in progress towards declarative datasets
Scope out ideas for a graph resolver that
derives sense hat data from the columns that
are present in a given dataset.
Also provide a downloader from GDrive to download
the raw assets.
The plan would be to use the graph resolver and
the downloader to create declarative scripts for
each prepared dataset.