Update asset upload script with schema validation
The CI build was failing because the metadata format
needed changing and the assets hand't been updated to
match. Took the opportunity to smarten up the `upload_assets`
script to make it easier to run, and to verify that uploaded
assets include the correct schema and no GPS tags. Some
integration tests have also been created to verify the same thing.